requirements.txt
gitdb2==2.0.0
GitPython==2.1.3
Jinja2==2.9.6
markdown2==2.3.3
MarkupSafe==1.0
python-dateutil==2.6.0
six==1.10.0
smmap2==2.0.1
webassets==0.12.1
wheel==0.24.0
